chore: changed domain name, timezone position
