+++ apps/web/src/lib/menu/menu-info.ts
@@ -179,12 +179,6 @@ export const MENU_INFO_MAP: Record<MenuId, MenuInfo> = Object.freeze({
         translationId: 'IAM.DOMAIN_SETTINGS.APPEARANCE',
     },
 
-    [MENU_ID.TIMEZONE_AND_LANGUAGE]: {
-        menuId: MENU_ID.TIMEZONE_AND_LANGUAGE,
-        routeName: PREFERENCE_ROUTE.DOMAIN_SETTINGS.TIMEZONE_AND_LANGUAGE._NAME,
-        translationId: 'IAM.DOMAIN_SETTINGS.TIMEZONE_AND_LANGUAGE',
-    },
-
     [MENU_ID.AUTO_DORMANCY_CONFIGURATION]: {
         menuId: MENU_ID.AUTO_DORMANCY_CONFIGURATION,
         routeName: PREFERENCE_ROUTE.DOMAIN_SETTINGS.AUTO_DORMANCY_CONFIGURATION._NAME,
