+++ apps/web/src/services/preference/pages/admin/AdminDomainSettingsDomainInformationPage.vue
@@ -2,60 +2,77 @@
 import { computed, reactive, watch } from 'vue';
 
 import {
-    PPaneLayout, PFieldTitle, PCopyButton, PTextInput, PButton,
+    PPaneLayout, PFieldTitle, PCopyButton, PButton, PSelectDropdown,
 } from '@spaceone/design-system';
+import type { SelectDropdownMenuItem } from '@spaceone/design-system/types/inputs/dropdown/select-dropdown/type';
+import { map } from 'lodash';
 
 import { store } from '@/store';
 import { i18n } from '@/translations';
 
 import { useDomainSettingsStore } from '@/store/domain-settings/domain-settings-store';
+import { languages, timezoneList } from '@/store/modules/user/config';
+import type { LanguageCode } from '@/store/modules/user/type';
 
 import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
-
-import DomainSettingsChangeAdminEmailModal
-    from '@/services/preference/components/DomainSettingsChangeAdminEmailModal.vue';
-
 
 const domainConfigStore = useDomainSettingsStore();
 const domainConfigGetters = domainConfigStore.getters;
 const storeState = reactive({
     domainId: computed<string>(() => store.state.domain.domainId),
     domainName: computed<string>(() => store.state.domain.name),
+    domainConfig: computed(() => store.state.domain.config),
 });
 const state = reactive({
-    displayName: undefined as string | undefined,
-    adminEmail: undefined as string | undefined,
-    isDisplayNameChanged: computed<boolean>(() => {
-        if (!state.displayName && !domainConfigGetters.displayName) return false;
-        return state.displayName !== domainConfigGetters.displayName;
+
+    isChanged: computed(() => {
+        if ([state.selectedTimezone, state.selectedLanguage,
+            domainConfigGetters.timezone, domainConfigGetters.language]
+            .every((d) => !d)) return false;
+        return (state.selectedTimezone !== domainConfigGetters.timezone)
+            || (state.selectedLanguage !== domainConfigGetters.language);
     }),
-    adminEmailChangeModalVisible: false,
+    selectedTimezone: undefined as string | undefined,
+    selectedLanguage: undefined as LanguageCode | undefined,
+    languageMenuList: computed<SelectDropdownMenuItem[]>(() => map(languages, (d, k) => ({
+        type: 'item', label: k === 'en' ? `${d} (default)` : d, name: k,
+    }))),
+    timezoneMenuList: computed<SelectDropdownMenuItem[]>(() => map(timezoneList, (d) => ({
+        type: 'item', label: d === 'UTC' ? `${d} (default)` : d, name: d,
+    }))),
+    config: computed(() => ({
+        ...storeState.domainConfig,
+        settings: {
+            timezone: state.selectedTimezone,
+            language: state.selectedLanguage,
+        },
+    })),
 });
 
 /* Event */
-const handleClickSaveDisplayName = async () => {
+const handleSaveChanges = async () => {
     try {
         await domainConfigStore.updateDomainSettings({
-            display_name: state.displayName,
+            timezone: state.selectedTimezone,
+            language: state.selectedLanguage,
         });
-        document.title = state.displayName || 'SpaceONE';
-        showSuccessMessage(i18n.t('IAM.DOMAIN_SETTINGS.ALT_S_UPDATE_DISPLAY_NAME'), '');
+        await store.commit('domain/setDomainConfigs', {
+            config: state.config,
+        });
+        showSuccessMessage(i18n.t('IAM.DOMAIN_SETTINGS.ALT_S_UPDATE_TIMEZONE_AND_LANGUAGE'), '');
     } catch (e) {
-        ErrorHandler.handleRequestError(e, i18n.t('IAM.DOMAIN_SETTINGS.ALT_E_UPDATE_DISPLAY_NAME'));
+        ErrorHandler.handleRequestError(e, i18n.t('IAM.DOMAIN_SETTINGS.ALT_E_UPDATE_TIMEZONE_AND_LANGUAGE'));
     }
 };
-// const handleClickChangeAdminEmail = () => {
-//     state.adminEmailChangeModalVisible = true;
-// };
 
 /* Watcher */
-watch(() => domainConfigGetters.displayName, (val) => {
-    state.displayName = val;
+watch(() => domainConfigGetters.timezone, (val) => {
+    if (val) state.selectedTimezone = val;
 }, { immediate: true });
-watch(() => domainConfigGetters.adminEmail, (val) => {
-    state.adminEmail = val;
+watch(() => domainConfigGetters.language, (val) => {
+    state.selectedLanguage = val;
 }, { immediate: true });
 </script>
 
@@ -77,59 +94,53 @@ watch(() => domainConfigGetters.adminEmail, (val) => {
                 </p>
             </div>
             <div class="field-wrapper">
-                <p-field-title :label="$t('IAM.DOMAIN_SETTINGS.DISPLAY_NAME')" />
-                <div class="input-wrapper">
-                    <p-text-input :value.sync="state.displayName" />
-                    <p-button class="ml-2"
-                              :disabled="!state.isDisplayNameChanged"
-                              @click="handleClickSaveDisplayName"
-                    >
-                        {{ $t('IAM.DOMAIN_SETTINGS.SAVE_CHANGE') }}
-                    </p-button>
-                </div>
+                <p-field-title :label="$t('IAM.DOMAIN_SETTINGS.TIMEZONE')" />
+                <p-select-dropdown :menu="state.timezoneMenuList"
+                                   :selected.sync="state.selectedTimezone"
+                                   :page-size="10"
+                                   show-delete-all-button
+                                   is-filterable
+                />
             </div>
-            <!--            <div class="field-wrapper">-->
-            <!--                <p-field-group :label="$t('IAM.DOMAIN_SETTINGS.ADMIN_EMAIL')"-->
-            <!--                               required-->
-            <!--                >-->
-            <!--                    <div class="input-wrapper">-->
-            <!--                        <p-text-input :value="state.adminEmail"-->
-            <!--                                      disabled-->
-            <!--                        />-->
-            <!--                        <p-button class="ml-2"-->
-            <!--                                  style-type="tertiary"-->
-            <!--                                  icon-left="ic_edit"-->
-            <!--                                  @click="handleClickChangeAdminEmail"-->
-            <!--                        >-->
-            <!--                            {{ $t('IAM.DOMAIN_SETTINGS.CHANGE') }}-->
-            <!--                        </p-button>-->
-            <!--                    </div>-->
-            <!--                </p-field-group>-->
-            <!--            </div>-->
+            <div class="field-wrapper">
+                <p-field-title :label="$t('IAM.DOMAIN_SETTINGS.LANGUAGE')" />
+                <p-select-dropdown :menu="state.languageMenuList"
+                                   :selected.sync="state.selectedLanguage"
+                                   :page-size="10"
+                                   show-delete-all-button
+                                   is-filterable
+                />
+            </div>
         </div>
-        <domain-settings-change-admin-email-modal :visible.sync="state.adminEmailChangeModalVisible" />
+        <p-button :disabled="!state.isChanged"
+                  class="save-button"
+                  @click="handleSaveChanges"
+        >
+            {{ $t('IAM.DOMAIN_SETTINGS.SAVE_CHANGES') }}
+        </p-button>
     </p-pane-layout>
 </template>
 
 <style lang="postcss" scoped>
 .admin-domain-settings-base-information-page {
     .content-wrapper {
-        display: flex;
-        flex-direction: column;
+        @apply flex flex-col;
+        width: 15rem;
         gap: 1.5rem;
         padding: 1rem;
         .field-wrapper {
             .input-wrapper {
-                display: flex;
-                align-items: center;
+                @apply flex items-center;
                 gap: 0.25rem;
                 padding-top: 0.25rem;
-                .p-text-input {
-                    max-width: 27rem;
-                    width: 100%;
-                }
+            }
+            .p-field-title {
+                padding-bottom: 0.25rem;
             }
         }
+    }
+    .save-button {
+        margin: 0.5rem 1rem 3rem;
     }
 }
 </style>
