+++ apps/web/src/services/preference/routes/route-constant.ts
@@ -6,7 +6,6 @@ export const PREFERENCE_ROUTE = {
         _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.DOMAIN_SETTINGS}`,
         DOMAIN_INFORMATION: { _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.DOMAIN_SETTINGS}.domain_information` },
         APPEARANCE: { _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.DOMAIN_SETTINGS}.appearance` },
-        TIMEZONE_AND_LANGUAGE: { _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.DOMAIN_SETTINGS}.timezone_and_language` },
         AUTO_DORMANCY_CONFIGURATION: { _NAME: `${MENU_ID.PREFERENCE}.${MENU_ID.DOMAIN_SETTINGS}.auto_dormancy_configuration` },
     },
     BOOKMARK: {
